Fix warning/error in debug assertion in device_uvector.hpp (#979)
Fixes #973

Authors:
  - Mark Harris (https://github.com/harrism)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Rong Ou (https://github.com/rongou)

URL: https://github.com/rapidsai/rmm/pull/979

+++ include/rmm/device_uvector.hpp
@@ -489,7 +489,8 @@ class device_uvector {
    */
   [[nodiscard]] std::int64_t ssize() const noexcept
   {
-    assert(size() < std::numeric_limits<int64_t>::max() && "Size overflows signed integer");
+    assert(size() < static_cast<std::size_t>(std::numeric_limits<int64_t>::max()) &&
+           "Size overflows signed integer");
     return static_cast<int64_t>(size());
   }
